Create compare route and update test
