+++ app/routes.ts
@@ -4,4 +4,5 @@ export default [
   index("routes/home.tsx"),
   route("welcome", "routes/welcome.tsx"),
   route("test", "routes/test.tsx"),
+  route("compare", "routes/compare.tsx"),
 ] satisfies RouteConfig;
